refactor : seperate roles (step3)
